Myt 275

+++ mType/vm/runtime/executors/TypeExecutor.cpp
@@ -1,4 +1,6 @@
 #include "TypeExecutor.hpp"
+#include <cstddef>
+#include <cstdint>
 #include "../utils/ErrorLocationHelper.hpp"
 #include "../utils/NullCheckUtils.hpp"
 #include "../utils/TypeArgResolution.hpp"
@@ -661,7 +663,7 @@ namespace vm::runtime
         }
         else if (value::isString(val)) {
             try {
-                return std::stoll(value::asString(val));
+                return static_cast<int64_t>(std::stoll(value::asString(val)));
             } catch (...) {
                 throwCastError("Cannot cast string to int: " + value::asString(val));
             }
@@ -669,7 +671,7 @@ namespace vm::runtime
         else if (value::isInternedString(val)) {
             const std::string& str = value::asInternedString(val).getString();
             try {
-                return std::stoll(str);
+                return static_cast<int64_t>(std::stoll(str));
             } catch (...) {
                 throwCastError("Cannot cast string to int: " + str);
             }
